refactor(model): split model.py into per-domain package + extract schema
Splits the 2065-line src/backend/klangk_backend/model.py into a model/
sub-package, one module per domain:

  __init__.py        re-exports every name (back-compat shim)
  _core.py           _Row/_CursorResult/_Connection, engine, get_db,
                     transaction(), _fetchone() (+ engine state)
  schema.py          init_db() + CREATE TABLE statements
  users.py           user CRUD, handles, groups, external IDs, agent user
  acl.py             ACL entries + ACTION/PRINCIPAL/SYSTEM constants
  workspaces.py      workspace CRUD + members + shares
  ports.py           port allocations + free-port discovery
  chat.py            chat messages + @mention resolution
  login_attempts.py  brute-force tracking
  tokens.py          JWT token blocklist
  invitations.py     invitation lifecycle

__init__.py re-exports all public (and the few private) names so every
existing `model.X` / `from .model import X` call site in api.py, auth.py,
main.py, oidc.py, acl.py and workspaces.py keeps working unchanged.

Only test plumbing that pokes module-level globals is retargeted to the
owning submodule: conftest.py sets engine state on model._core, and
test_model.py monkeypatches model.ports._port_in_use / reads
model._core.DB_PATH.

Backend suite: 1820 passed, 100% coverage.

Refs #968.
