+++ src/backend/klangk_backend/model/__init__.py
@@ -1,0 +1,283 @@
+"""Data-access layer.
+
+Historically all database access — users, workspaces, ACL, chat, ports,
+login attempts, and the schema — lived in a single ~2000-line
+``model.py``.  That module has been split into per-domain submodules
+(``_core``, ``schema``, ``users``, ``acl``, ``workspaces``, ``ports``,
+``chat``, ``login_attempts``, ``tokens``, ``invitations``).
+
+This package re-exports every public (and the few private) names from
+those submodules so existing call sites keep working unchanged, e.g.::
+
+    from . import model
+    model.create_user(...)
+    from .model import ACTION_ALLOW
+
+Only the handful of tests that poke at module-level globals need to
+target the owning submodule directly (``model._core`` for engine state,
+``model.ports`` for the port-in-use probe).
+"""
+
+from ._core import (
+    DB_PATH,
+    _Connection,
+    _CursorResult,
+    _Row,
+    _data_dir,
+    _engine,
+    _ensure_engine,
+    _fetchone,
+    _make_engine,
+    dispose_engine,
+    get_db,
+    logger,
+    transaction,
+)
+from .schema import init_db
+from .users import (
+    AGENT_USER_ID,
+    _ADMIN_USER_SORT_COLUMNS,
+    _HANDLE_RE,
+    _MAX_HANDLE_LEN,
+    _RESERVED_HANDLES,
+    _agent_user_cache,
+    _backfill_handles,
+    _hash_fallback_handle,
+    _unique_handle,
+    add_user_to_group,
+    agent_email,
+    agent_handle,
+    clear_agent_cache,
+    create_group,
+    create_user,
+    delete_group,
+    delete_user,
+    derive_handle,
+    get_agent_user,
+    get_group_by_id,
+    get_group_by_name,
+    get_group_members,
+    get_user_by_email,
+    get_user_by_external_id,
+    get_user_by_handle,
+    get_user_by_id,
+    get_user_group_ids,
+    get_user_groups,
+    get_user_handle,
+    get_user_oidc_sync_group_ids,
+    link_oidc_identity,
+    list_groups,
+    list_users,
+    remove_user_from_group,
+    search_users,
+    set_user_handle,
+    update_email,
+    update_group,
+    update_password,
+    validate_handle,
+    verify_user,
+)
+from .acl import (
+    ACTION_ALLOW,
+    ACTION_DENY,
+    PRINCIPAL_GROUP,
+    PRINCIPAL_SYSTEM,
+    PRINCIPAL_USER,
+    SYSTEM_AUTHENTICATED,
+    SYSTEM_EVERYONE,
+    _row_to_acl_entry,
+    add_acl_entry,
+    delete_acl_entries_for_resource,
+    get_acl_entries,
+    get_acl_entries_by_principal_group,
+    get_acl_entries_by_principal_user,
+    get_acl_entries_map,
+    get_acl_entries_resolved,
+    get_acl_tree_summary,
+    replace_acl_entries,
+)
+from .workspaces import (
+    _DEFAULT_PORTS_PER_WORKSPACE,
+    _SORT_COLUMNS,
+    _sort_order_clause,
+    create_workspace,
+    delete_workspace,
+    get_user_workspaces_with_containers,
+    get_workspace,
+    get_workspace_by_id,
+    get_workspace_members,
+    list_shared_workspaces,
+    list_workspaces,
+    update_workspace,
+    update_workspace_container,
+)
+from .ports import (
+    MAX_PORT,
+    _port_in_use,
+    _scan_free_ports,
+    add_port_allocations,
+    find_and_allocate_ports,
+    get_all_allocated_ports,
+    get_workspace_ports,
+    remove_port_allocations,
+)
+from .chat import (
+    MSG_AGENT,
+    MSG_SYSTEM,
+    MSG_USER,
+    _MENTION_RE,
+    add_chat_message,
+    delete_chat_message,
+    get_chat_messages,
+    get_chat_messages_before,
+    parse_mentions,
+)
+from .login_attempts import (
+    clear_login_attempts,
+    get_login_attempt_info,
+    record_failed_login,
+    set_login_lockout,
+)
+from .tokens import (
+    blocklist_token,
+    get_refreshed_token,
+    is_token_blocklisted,
+)
+from .invitations import (
+    _ADMIN_INVITATION_SORT_COLUMNS,
+    create_invitation,
+    get_invitation,
+    get_pending_invitation_by_email,
+    list_invitations,
+    mark_invitation_accepted,
+    revoke_invitation,
+)
+
+__all__ = (
+    # _core
+    "DB_PATH",
+    "_Connection",
+    "_CursorResult",
+    "_Row",
+    "_data_dir",
+    "_engine",
+    "_ensure_engine",
+    "_fetchone",
+    "_make_engine",
+    "dispose_engine",
+    "get_db",
+    "logger",
+    "transaction",
+    # schema
+    "init_db",
+    # users
+    "AGENT_USER_ID",
+    "_ADMIN_USER_SORT_COLUMNS",
+    "_HANDLE_RE",
+    "_MAX_HANDLE_LEN",
+    "_RESERVED_HANDLES",
+    "_agent_user_cache",
+    "_backfill_handles",
+    "_hash_fallback_handle",
+    "_unique_handle",
+    "add_user_to_group",
+    "agent_email",
+    "agent_handle",
+    "clear_agent_cache",
+    "create_group",
+    "create_user",
+    "delete_group",
+    "delete_user",
+    "derive_handle",
+    "get_agent_user",
+    "get_group_by_id",
+    "get_group_by_name",
+    "get_group_members",
+    "get_user_by_email",
+    "get_user_by_external_id",
+    "get_user_by_handle",
+    "get_user_by_id",
+    "get_user_group_ids",
+    "get_user_groups",
+    "get_user_handle",
+    "get_user_oidc_sync_group_ids",
+    "link_oidc_identity",
+    "list_groups",
+    "list_users",
+    "remove_user_from_group",
+    "search_users",
+    "set_user_handle",
+    "update_email",
+    "update_group",
+    "update_password",
+    "validate_handle",
+    "verify_user",
+    # acl
+    "ACTION_ALLOW",
+    "ACTION_DENY",
+    "PRINCIPAL_GROUP",
+    "PRINCIPAL_SYSTEM",
+    "PRINCIPAL_USER",
+    "SYSTEM_AUTHENTICATED",
+    "SYSTEM_EVERYONE",
+    "_row_to_acl_entry",
+    "add_acl_entry",
+    "delete_acl_entries_for_resource",
+    "get_acl_entries",
+    "get_acl_entries_by_principal_group",
+    "get_acl_entries_by_principal_user",
+    "get_acl_entries_map",
+    "get_acl_entries_resolved",
+    "get_acl_tree_summary",
+    "replace_acl_entries",
+    # workspaces
+    "_DEFAULT_PORTS_PER_WORKSPACE",
+    "_SORT_COLUMNS",
+    "_sort_order_clause",
+    "create_workspace",
+    "delete_workspace",
+    "get_user_workspaces_with_containers",
+    "get_workspace",
+    "get_workspace_by_id",
+    "get_workspace_members",
+    "list_shared_workspaces",
+    "list_workspaces",
+    "update_workspace",
+    "update_workspace_container",
+    # ports
+    "MAX_PORT",
+    "_port_in_use",
+    "_scan_free_ports",
+    "add_port_allocations",
+    "find_and_allocate_ports",
+    "get_all_allocated_ports",
+    "get_workspace_ports",
+    "remove_port_allocations",
+    # chat
+    "MSG_AGENT",
+    "MSG_SYSTEM",
+    "MSG_USER",
+    "_MENTION_RE",
+    "add_chat_message",
+    "delete_chat_message",
+    "get_chat_messages",
+    "get_chat_messages_before",
+    "parse_mentions",
+    # login_attempts
+    "clear_login_attempts",
+    "get_login_attempt_info",
+    "record_failed_login",
+    "set_login_lockout",
+    # tokens
+    "blocklist_token",
+    "get_refreshed_token",
+    "is_token_blocklisted",
+    # invitations
+    "_ADMIN_INVITATION_SORT_COLUMNS",
+    "create_invitation",
+    "get_invitation",
+    "get_pending_invitation_by_email",
+    "list_invitations",
+    "mark_invitation_accepted",
+    "revoke_invitation",
+)
