+++ src/backend/tests/conftest.py
@@ -31,14 +31,18 @@ def temp_data_dir(tmp_path, monkeypatch):
     """Point KLANGK_DATA_DIR to a temp directory for each test."""
     monkeypatch.setenv("KLANGK_DATA_DIR", str(tmp_path))
     monkeypatch.delenv("KLANGK_IMAGE_PULL_POLICY", raising=False)
-    # Re-import to pick up the new env var
+    # Re-import to pick up the new env var.  The engine state lives on
+    # ``model._core`` (where ``_ensure_engine`` reads it), so the globals
+    # must be rebound there — setting them on the package would not be
+    # seen by the core helpers.
     import klangk_backend.model as us
+    import klangk_backend.model._core as us_core
     import klangk_backend.workspaces as wm
 
-    us._data_dir = tmp_path
-    us.DB_PATH = tmp_path / "klangk.db"
+    us_core._data_dir = tmp_path
+    us_core.DB_PATH = tmp_path / "klangk.db"
     # Reset the SQLAlchemy engine so it reconnects to the new DB path.
-    us._engine = None
+    us_core._engine = None
     wm._data_dir = tmp_path
     wm.WORKSPACES_ROOT = tmp_path / "workspaces"
     # Clear agent caches so each test starts fresh.
